+++ backend/src/main/java/com/maarcus/backend/controller/ColorController.java
@@ -3,67 +3,99 @@ package com.maarcus.backend.controller;
 import com.maarcus.backend.exception.category.CategoryNotFoundException;
 import com.maarcus.backend.model.Color;
 import com.maarcus.backend.service.ColorService;
-import com.maarcus.backend.model.StandardResponse;
+import com.maarcus.backend.payload.response.StandardResponse;
 import com.maarcus.backend.exception.color.ColorNotFoundException;
+import com.maarcus.backend.utils.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/colors")
+@RequestMapping(path = "/api/colors")
 public class ColorController {
   
   private final ColorService colorService;
+  private final ResponseUtil responseUtil;
   
   @Autowired
-  public ColorController(ColorService colorService) {
+  public ColorController(ColorService colorService, ResponseUtil responseUtil) {
     this.colorService = colorService;
+	  this.responseUtil = responseUtil;
   }
   
-  @GetMapping("/all")
+  @GetMapping(path = "/all")
   public ResponseEntity<StandardResponse<List<Color>>> getAllColors() {
-
-    return ResponseEntity.ok(new StandardResponse<>(HttpStatus.OK, "Colors retrieved successfully", colorService.getAllColors()));
-  }
-  
-  @GetMapping("/get/{id}")
-  public ResponseEntity<StandardResponse<Optional<Color>>> getColor(@PathVariable Long id) {
-    try {
-      return ResponseEntity.ok(new StandardResponse<>(HttpStatus.OK, "Color retrieved successfully", colorService.getColor(id)));
-    } catch (ColorNotFoundException ex) {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new StandardResponse<>(HttpStatus.NOT_FOUND, ex.getMessage(), null));
+    List<Color> colors = colorService.getAllColors();
+    
+    if (colors.isEmpty()) {
+      return responseUtil.buildErrorResponse(HttpStatus.NOT_FOUND, "No categories found");
     }
+    
+    return responseUtil.buildSuccessResponse(HttpStatus.OK, "Categories retrieved successfully", colors);
   }
   
-  @PostMapping("/add")
-  public ResponseEntity<StandardResponse<Optional<Color>>> addColor(@RequestBody Color color) {
-    try {
-      return ResponseEntity.status(HttpStatus.CREATED).body(new StandardResponse<>(HttpStatus.CREATED, "Color added successfully", colorService.addColor(color)));
-    } catch (Exception ex) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new StandardResponse<>(HttpStatus.BAD_REQUEST, ex.getMessage(), null));
+  @GetMapping(path = "/get/{id}")
+  public ResponseEntity<StandardResponse<Color>> getColor(@PathVariable Long id) {
+    
+    if (id == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: id");
     }
+    
+    Color color = colorService.getColor(id).orElseThrow(() -> new ColorNotFoundException(id));
+    
+    return responseUtil.buildSuccessResponse(HttpStatus.OK, "Color retrieved successfully", color);
   }
   
-  @PutMapping("/update/{id}")
+  @PostMapping(path = "/add")
+  public ResponseEntity<StandardResponse<Color>> addColor(@RequestBody Color color) {
+    if (color.getName() == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: name");
+    }
+    
+    if (color.getValue() == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: value");
+    }
+    
+    Color createdColor = colorService.addColor(color);
+    return responseUtil.buildSuccessResponse(HttpStatus.CREATED, "Color created successfully", createdColor);
+  }
+  
+  @PutMapping(path = "/update/{id}")
   public ResponseEntity<StandardResponse<Color>> updateColor(@PathVariable Long id, @RequestBody Color color) {
+    if (id == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: id");
+    }
+    
+    if (color.getName() == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: name");
+    }
+    
+    if (color.getValue() == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: value");
+    }
+    
     try {
-      return ResponseEntity.ok(new StandardResponse<>(HttpStatus.OK, "Color updated successfully", colorService.updateColor(id, color)));
-    } catch (ColorNotFoundException ex) {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new StandardResponse<>(HttpStatus.NOT_FOUND, ex.getMessage(), null));
+      Color updatedColor = colorService.updateColor(id, color);
+      return responseUtil.buildSuccessResponse(HttpStatus.OK, "Color updated successfully", updatedColor);
+    } catch (CategoryNotFoundException e) {
+      return responseUtil.buildErrorResponse(HttpStatus.NOT_FOUND, e.getMessage());
     }
   }
   
-  @DeleteMapping("/delete/{id}")
+  @DeleteMapping(path = "/delete/{id}")
   public ResponseEntity<StandardResponse<Void>> deleteColor(@PathVariable Long id) {
+    if (id == null) {
+      return responseUtil.buildErrorResponse(HttpStatus.BAD_REQUEST, "Missing required field: id");
+    }
+    
     try {
       colorService.deleteColor(id);
-      return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new StandardResponse<>(HttpStatus.NO_CONTENT, "Color deleted successfully", null));
-    } catch (CategoryNotFoundException ex) {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new StandardResponse<>(HttpStatus.NOT_FOUND, ex.getMessage(), null));
+      return responseUtil.buildSuccessResponse(HttpStatus.NO_CONTENT, "Color deleted successfully", null);
+    } catch (CategoryNotFoundException e) {
+      return responseUtil.buildErrorResponse(HttpStatus.NOT_FOUND, e.getMessage());
     }
   }
 }
